Testeando Efectos
Post Action aun no testeado
Efecto de Damage Hecho y testeado

+++ Assets/Scripts/Compiler/Parser/ASTNodes/Terminals/NumberNode.cs
@@ -4,9 +4,9 @@ namespace Compiler
 {
     public class NumberNode : TerminalNodes
     {
-        public double Value;
+        public int Value;
 
-        public NumberNode (double value)
+        public NumberNode (int value)
         {
             this.Value = value;
         }
